[pam_misc] Split header from code
